Moves Theme to its own file Moves everything out of the GameWindow and into its own GameView page.
