+++ Sudoku/GameWindow.xaml
@@ -7,158 +7,6 @@
 		mc:Ignorable="d"
 		Title="GameWindow"
 		Height="500"
-		Width="800"
-		KeyDown="Window_KeyDown">
-	<!-- The main grid, left and right panes. -->
-	<Grid ShowGridLines="True">
-		<Grid.RowDefinitions>
-			<RowDefinition Height="*" />
-		</Grid.RowDefinitions>
-		<Grid.ColumnDefinitions>
-			<ColumnDefinition Width="*" />
-			<ColumnDefinition Width="*" />
-		</Grid.ColumnDefinitions>
-		<!-- The new game button -->
-		<Button Content="New Game"
-				Grid.Column="0"
-				HorizontalAlignment="Left"
-				Margin="10,10,0,0"
-				VerticalAlignment="Top"
-				Click="New_Game" />
-		<!-- The sudoku board -->
-		<Grid Grid.Column="0"
-			  ShowGridLines="True"
-			  Name="gridView"
-			  Width="300"
-			  Height="300"
-			  HorizontalAlignment="Center">
-			<Grid.RowDefinitions>
-				<RowDefinition />
-				<RowDefinition />
-				<RowDefinition />
-				<RowDefinition />
-				<RowDefinition />
-				<RowDefinition />
-				<RowDefinition />
-				<RowDefinition />
-				<RowDefinition />
-			</Grid.RowDefinitions>
-			<Grid.ColumnDefinitions>
-				<ColumnDefinition />
-				<ColumnDefinition />
-				<ColumnDefinition />
-				<ColumnDefinition />
-				<ColumnDefinition />
-				<ColumnDefinition />
-				<ColumnDefinition />
-				<ColumnDefinition />
-				<ColumnDefinition />
-			</Grid.ColumnDefinitions>
-		</Grid>
-		<!-- StackPanel used to organize the right panel -->
-		<StackPanel Grid.Column="1">
-			<Grid ShowGridLines="True"
-				  Width="200"
-				  Height="200"
-				  Margin="25">
-				<Grid.RowDefinitions>
-					<RowDefinition />
-					<RowDefinition />
-					<RowDefinition />
-				</Grid.RowDefinitions>
-				<Grid.ColumnDefinitions>
-					<ColumnDefinition />
-					<ColumnDefinition />
-					<ColumnDefinition />
-				</Grid.ColumnDefinitions>
-				<Button Grid.Row="0"
-						Grid.Column="0"
-						Content="1"
-						FontSize="30"
-						Background="White"
-						Click="KeyPad" />
-				<Button Grid.Row="0"
-						Grid.Column="1"
-						Content="2"
-						FontSize="30"
-						Background="White"
-						Click="KeyPad" />
-				<Button Grid.Row="0"
-						Grid.Column="2"
-						Content="3"
-						FontSize="30"
-						Background="White"
-						Click="KeyPad" />
-				<Button Grid.Row="1"
-						Grid.Column="0"
-						Content="4"
-						FontSize="30"
-						Background="White"
-						Click="KeyPad" />
-				<Button Grid.Row="1"
-						Grid.Column="1"
-						Content="5"
-						FontSize="30"
-						Background="White" />
-				<Button Grid.Row="1"
-						Grid.Column="2"
-						Content="6"
-						FontSize="30"
-						Background="White" />
-				<Button Grid.Row="2"
-						Grid.Column="0"
-						Content="7"
-						FontSize="30"
-						Background="White" />
-				<Button Grid.Row="2"
-						Grid.Column="1"
-						Content="8"
-						FontSize="30"
-						Background="White" />
-				<Button Grid.Row="2"
-						Grid.Column="2"
-						Content="9"
-						FontSize="30"
-						Background="White" />
-			</Grid>
-			<Grid Width="300">
-				<Grid.RowDefinitions>
-					<RowDefinition />
-					<RowDefinition />
-				</Grid.RowDefinitions>
-				<Grid.ColumnDefinitions>
-					<ColumnDefinition />
-					<ColumnDefinition />
-				</Grid.ColumnDefinitions>
-				<Button Width="100"
-						Height="50"
-						Margin="10"
-						Grid.Row="0"
-						Grid.Column="0"
-						Content="Note" />
-				<Button Width="100"
-						Height="50"
-						Margin="10"
-						Grid.Row="0"
-						Grid.Column="1"
-						Content="Hint" />
-				<Button Width="100"
-						Height="50"
-						Margin="25"
-						Grid.Row="1"
-						Grid.Column="0"
-						Content="Undo" />
-				<Button Width="100"
-						Height="50"
-						Margin="25"
-						Grid.Row="1"
-						Grid.Column="1"
-						Content="Erase" />
-			</Grid>
-			<Button Width="100"
-					Height="50"
-					Content="Reset"
-					Click="Reset_Board" />
-		</StackPanel>
-	</Grid>
+		Width="800">
+	<Frame Name="mainFrame" Source="GameView.xaml"/>
 </Window>
